Move rest to base/rest
Caleydo/cohort#693
